增加统计实现

+++ core/src/main/java/cn/hz/ddbm/pc/core/support/StatisticsSupport.java
@@ -12,5 +12,5 @@ package cn.hz.ddbm.pc.core.support;
 public interface StatisticsSupport {
     void increment(String windows);
 
-    Integer get(String windows);
+    Long get(String windows);
 }
