+++ example/src/main/java/cn/hz/ddbm/pc/test/support/MetricsTemplateMock.java
@@ -5,17 +5,18 @@ import cn.hz.ddbm.pc.core.support.StatisticsSupport;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class MetricsTemplateMock implements StatisticsSupport {
-    Map<String, AtomicInteger> map = new HashMap<>();
+    Map<String, AtomicLong> map = new HashMap<>();
 
     @Override
     public void increment(String windows) {
-        map.computeIfAbsent(windows, s -> new AtomicInteger(0)).incrementAndGet();
+        map.computeIfAbsent(windows, s -> new AtomicLong(0)).incrementAndGet();
     }
 
     @Override
-    public Integer get(String windows) {
-        return map.computeIfAbsent(windows, s -> new AtomicInteger(0)).get();
+    public Long get(String windows) {
+        return map.computeIfAbsent(windows, s -> new AtomicLong(0)).get();
     }
 }
